Complete overhaul of process tracking system Instead of holding a process state for each process event, all process info is held in one structure. Process info that changes throughout time is represented in a binary tree indexed by the start time of that piece of info. Process info dissection displays all process info across its lifespan, while annotating relevance of time-dependent pieces of info, and referencing events in the process' lifecycle.
In order to avoid messy and innacurate logic regarding the order in
which process events are encountered (and how to decide when info
becomes valid/invalid), a requirement was added for
process tracking - the capture must be ordered chronologically.
A capture can be ordered using the `reordercap` utility.

Right now, only fork events contribute to process tracking.
Other events will be added next.